% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square.R
\name{image_square}
\alias{image_square}
\title{Crop to Square}
\usage{
image_square(image, gravity = "center")
}
\arguments{
\item{image}{magick image object returned by \code{\link[magick:image_read]{image_read()}} or \code{\link[magick:image_graph]{image_graph()}}}

\item{gravity}{string with
\href{https://www.imagemagick.org/Magick++/Enumerations.html#GravityType}{gravity}
value from \link[magick]{gravity_types}.}
}
\value{
Magick image object.
}
\description{
Crop to Square
}
\examples{
image_square(magick::logo)
}
