% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radon-package.R
\docType{package}
\name{radon-package}
\alias{radon}
\alias{radon-package}
\title{radon: Radon Transform for Magick Image Objects}
\description{
Apply the Radon transform to 'magick' image objects. Radon, Johann (1917) \doi{10.1109/TMI.1986.4307775}.
}
\keyword{internal}
